requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
idna==3.3
Pillow==8.4.0
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
